+++ version_history.md
@@ -2,6 +2,7 @@
 
 ## 2.1.5
 - Reworked marble reflection code, making it more performant and beautiful :)
+- Added support for a weird invisible MBU material
 
 ## 2.1.4
 - Fixed old replays again